move fnType method to DefinedType
